+++ src/java/org/apache/cassandra/service/pager/AbstractQueryPager.java
@@ -110,7 +110,7 @@ abstract class AbstractQueryPager<T extends ReadQuery> implements QueryPager
     private class UnfilteredPager extends Pager<Unfiltered>
     {
 
-        private UnfilteredPager(DataLimits pageLimits, int nowInSec)
+        private UnfilteredPager(DataLimits pageLimits, long nowInSec)
         {
             super(pageLimits, nowInSec);
         }
@@ -124,7 +124,7 @@ abstract class AbstractQueryPager<T extends ReadQuery> implements QueryPager
     private class RowPager extends Pager<Row>
     {
 
-        private RowPager(DataLimits pageLimits, int nowInSec)
+        private RowPager(DataLimits pageLimits, long nowInSec)
         {
             super(pageLimits, nowInSec);
         }
@@ -143,7 +143,7 @@ abstract class AbstractQueryPager<T extends ReadQuery> implements QueryPager
         private Row lastRow;
         private boolean isFirstPartition = true;
 
-        private Pager(DataLimits pageLimits, int nowInSec)
+        private Pager(DataLimits pageLimits, long nowInSec)
         {
             this.counter = pageLimits.newCounter(nowInSec, true, query.selectsFullPartition(), enforceStrictLiveness);
             this.pageLimits = pageLimits;
